:lipstick: fix line-height with 'leading-normal'
Remove custom padding from inside of Title component, and make its
text-height consistent across sizes.

Any custom padding that needs to be added should be passed to the Title
component in the classNames param.

Also modified the padding passed to Title in places where the padding
was not even arount the text. (As a rule of thumb, simpler padding rules
look better).